+++ src/propertyFields/termPicker/PropertyFieldTermPicker.ts
@@ -28,6 +28,7 @@ class PropertyFieldTermPickerBuilder implements IPropertyPaneField<IPropertyFiel
   private limitByTermsetNameOrID: string = null;
   private panelTitle: string;
   private hideTermStoreName: boolean;
+  private isTermSetSelectable: boolean;
 
   public onPropertyChange(propertyPath: string, oldValue: any, newValue: any): void { }
   private customProperties: any;
@@ -55,6 +56,7 @@ class PropertyFieldTermPickerBuilder implements IPropertyPaneField<IPropertyFiel
     this.limitByGroupNameOrID = _properties.limitByGroupNameOrID;
     this.limitByTermsetNameOrID = _properties.limitByTermsetNameOrID;
     this.hideTermStoreName = _properties.hideTermStoreName;
+    this.isTermSetSelectable = _properties.isTermSetSelectable;
 
     if (_properties.disabled === true) {
       this.disabled = _properties.disabled;
@@ -88,6 +90,7 @@ class PropertyFieldTermPickerBuilder implements IPropertyPaneField<IPropertyFiel
       limitByGroupNameOrID: this.limitByGroupNameOrID,
       limitByTermsetNameOrID: this.limitByTermsetNameOrID,
       hideTermStoreName: this.hideTermStoreName,
+      isTermSetSelectable: this.isTermSetSelectable,
       context: this.context,
       onDispose: this.dispose,
       onRender: this.render,